requirements.txt
# web側
fastapi
websockets
uvicorn
jinja2

# webに送る側
pyserial
websocket-client